aredn-compat: height can be float

+++ internal/server/api/apimodels/aredn_compat.go
@@ -394,12 +394,12 @@ type MeshRF1Point7 struct {
 
 type MeshRF1Point13 struct {
 	MeshRF1Point7
-	Azimuth    int    `json:"azimuth,string,omitempty"`
-	Elevation  int    `json:"elevation,string,omitempty"`
-	Height     int    `json:"height,string,omitempty"`
-	Antenna    any    `json:"antenna,omitempty"`
-	AntennaAux any    `json:"antenna_aux,omitempty"`
-	Mode       string `json:"mode,omitempty"`
+	Azimuth    int     `json:"azimuth,string,omitempty"`
+	Elevation  int     `json:"elevation,string,omitempty"`
+	Height     float64 `json:"height,string,omitempty"`
+	Antenna    any     `json:"antenna,omitempty"`
+	AntennaAux any     `json:"antenna_aux,omitempty"`
+	Mode       string  `json:"mode,omitempty"`
 }
 
 type MeshRF2Point0 struct {
